Starting work on PTypes
